code cleanup

+++ app/src/main/java/com/camo/kripto/remote/model/ExchangeRates.kt
@@ -1,12 +1,14 @@
 package com.camo.kripto.remote.model
 
+import java.math.BigDecimal
+
 data class ExchangeRates(
     val rates: Map<String, NVUT>
 ) {
     data class NVUT(
         val name: String,
         val unit: String,
-        val value: Double,
+        val value: BigDecimal,
         val type: String
     )
 }
