+++ app/src/main/java/com/camo/kripto/remote/model/Exchanges.kt
@@ -1,5 +1,7 @@
 package com.camo.kripto.remote.model
 
+import java.math.BigDecimal
+
 class Exchanges : ArrayList<Exchanges.ExchangesItem>(){
     data class ExchangesItem(
         val id: String,
@@ -12,7 +14,7 @@ class Exchanges : ArrayList<Exchanges.ExchangesItem>(){
         val has_trading_incentive: Boolean,
         val trust_score: Int,
         val trust_score_rank: Int,
-        val trade_volume_24h_btc: Double,
-        val trade_volume_24h_btc_normalized: Double
+        val trade_volume_24h_btc: BigDecimal,
+        val trade_volume_24h_btc_normalized: BigDecimal
     )
 }
