+++ app/src/main/java/com/camo/kripto/remote/model/MarketChart.kt
@@ -1,7 +1,9 @@
 package com.camo.kripto.remote.model
 
+import java.math.BigDecimal
+
 data class MarketChart(
-    val prices: List<List<Double>>,
-    val market_caps: List<List<Double>>,
-    val total_volumes: List<List<Double>>
+    val prices: List<List<BigDecimal>>,
+    val market_caps: List<List<BigDecimal>>,
+    val total_volumes: List<List<BigDecimal>>
 )
